Added JAVA source code for PPM-DP.
The code contains abstraction layers and tools for debugging
and visualisation.  A runnable demo version of PPM-DP for
regular files is contained in "PPMDP.java" and can be run
from the command line.

The code can be compiled with "javac PPMDP.java", and run
with "java PPMDP".  Try "java PPMDP --help" for some basic
usage information.